[IMP]stock product_expiry: move group_expiration_date_on_delivery_slip
This moves group_expiration_date_on_delivery_slip from module stock , to module product_expiry

Project : RD for dummies
Task : [Inventory] Add SN/LN on Delivery Order (#1832574)
