+++ addons/product_expiry/__manifest__.py
@@ -17,8 +17,15 @@ Following dates can be tracked:
 
 Also implements the removal strategy First Expiry First Out (FEFO) widely used, for example, in food industries.
 """,
-    'data': ['views/production_lot_views.xml',
-             'views/product_template_views.xml',
-             'views/stock_quant_views.xml',
-             'data/product_expiry_data.xml'],
+    'data': [
+        'security/stock_security.xml',
+
+        'views/production_lot_views.xml',
+        'views/product_template_views.xml',
+        'views/res_config_settings_views.xml',
+        'views/res_users_views.xml',
+        'views/stock_quant_views.xml',
+
+        'data/product_expiry_data.xml',
+    ],
 }
