+++ addons/product_expiry/models/__init__.py
@@ -3,4 +3,6 @@
 
 from . import production_lot
 from . import product_product
+from . import res_config_settings
+from . import res_users
 from . import stock_quant
