+++ addons/product_expiry/security/stock_security.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+<data noupdate="0">
+
+    <!-- move this start -->
+    <record id="group_expiration_date_on_delivery_slip" model="res.groups">
+        <field name="name">Display Expiration date in Delivery Slips</field>
+        <field name="category_id" ref="base.module_category_hidden"/>
+    </record>
+    <!-- move this end -->
+
+</data>
+</odoo>
