+++ addons/product_expiry/views/res_config_settings_views.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+        <record id="res_config_settings_view_form" model="ir.ui.view">
+            <field name="name">res.config.settings.view.form.inherit.stock</field>
+            <field name="model">res.config.settings</field>
+            <field name="priority" eval="30"/>
+            <field name="inherit_id" ref="stock.res_config_settings_view_form" />
+            <field name="arch" type="xml">
+                <xpath expr="//div[@id='group_lot_on_delivery_slip']" position="after">
+
+                    <!-- move this start -->
+                    <div class="col-xs-12 col-md-6 o_setting_box" attrs="{'invisible': ['|', ('module_product_expiry', '=', False), ('group_lot_on_delivery_slip', '=', False)]}">
+                        <div class="o_setting_left_pane">
+                            <field name="group_expiration_date_on_delivery_slip"/>
+                        </div>
+                        <div class="o_setting_right_pane">
+                            <label for="group_expiration_date_on_delivery_slip"/>
+                            <div class="text-muted">
+                                Expiration dates will appear on the delivery slip
+                            </div>
+                        </div>
+                    </div>
+                    <!-- move this end -->
+
+                </xpath>
+            </field>
+        </record>
+    </data>
+</odoo>
