+++ addons/product_expiry/views/res_users_views.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<odoo>
+    <record id="res_users_view_form" model="ir.ui.view">
+        <field name="name">res.users.form.stock.inherit</field>
+        <field name="model">res.users</field>
+        <field name="inherit_id" ref="base.view_users_form"/>
+        <field name="priority">15</field>
+        <field name="arch" type="xml">
+            <group name="technical" position="inside">
+                <!-- move this start -->
+                <label for="has_group_expiration_date_on_delivery_slip" colspan="3"/>
+                <div>
+                    <field name='has_group_expiration_date_on_delivery_slip' colspan="1" class="oe_inline"/>
+                </div>
+                <!-- move this end -->
+            </group>
+        </field>
+    </record>
+</odoo>
