+++ addons/stock/models/res_users.py
@@ -27,11 +27,6 @@ class ResUsers(models.Model):
         compute='_compute_groups_id', inverse='_inverse_groups_id',
         group_xml_id='stock.group_lot_on_delivery_slip')
 
-    has_group_expiration_date_on_delivery_slip = fields.Boolean(
-        'Display Expiration Dates on Delivery Slip',
-        compute='_compute_groups_id', inverse='_inverse_groups_id',
-        group_xml_id='stock.group_expiration_date_on_delivery_slip')
-
     has_group_tracking_lot = fields.Boolean(
         'Manage Packages',
         compute='_compute_groups_id', inverse='_inverse_groups_id',
