+++ addons/stock/security/stock_security.xml
@@ -31,11 +31,6 @@
         <field name="category_id" ref="base.module_category_hidden"/>
     </record>
 
-    <record id="group_expiration_date_on_delivery_slip" model="res.groups">
-        <field name="name">Display Expiration date in Delivery Slips</field>
-        <field name="category_id" ref="base.module_category_hidden"/>
-    </record>
-
     <record id="group_tracking_lot" model="res.groups">
         <field name="name">Manage Packages</field>
     </record>
