+++ addons/stock/views/res_users_views.xml
@@ -26,10 +26,6 @@
                 <div>
                     <field name='has_group_lot_on_delivery_slip' colspan="1" class="oe_inline"/>
                 </div>
-                <label for="has_group_expiration_date_on_delivery_slip" colspan="3"/>
-                <div>
-                    <field name='has_group_expiration_date_on_delivery_slip' colspan="1" class="oe_inline"/>
-                </div>
                 <label for="has_group_tracking_lot" colspan="3"/>
                 <div>
                     <field name='has_group_tracking_lot' colspan="1" class="oe_inline"/>
